Rework HNil to properly work with 2.12's exhaustivity checks.
We were using a singleton instnace of a final HNil class, but Scala 2.12
correctly detects a match that only mentions the singleton as incomplete in
that case, because someone else could instantiate the HNil class and use it.
We now use a case object, and provide a type alias for its singleton type so
that implementing code remains unchanged.